Bugfix on main.py command pair symbol argument Bugfix on Price type to store price as decimals

+++ models/price.py
@@ -3,7 +3,7 @@ from sqlalchemy import create_engine
 from sqlalchemy.ext.declarative import declarative_base
 from sqlalchemy.orm import sessionmaker
 from sqlalchemy.dialects.postgresql import UUID
-from sqlalchemy import Column, String, Float, DateTime
+from sqlalchemy import Column, String, Float, DateTime, DECIMAL
 import uuid
 import datetime
 
@@ -16,9 +16,9 @@ class Price(base):
 
     uuid = Column(UUID(as_uuid=True), primary_key=True, default=uuid.uuid4, unique=True)
     pair = Column(String, nullable=False)
-    curr = Column(Float, nullable=False)
-    lowest = Column(Float, nullable=False)
-    highest = Column(Float, nullable=False)
+    curr = Column(DECIMAL(16, 8), nullable=False)
+    lowest = Column(DECIMAL(16, 8), nullable=False)
+    highest = Column(DECIMAL(16, 8), nullable=False)
     created = Column(DateTime, default=datetime.datetime.utcnow)
